test: consolidate generated dependency-pin tests into one suite
CodeRabbit's test generation ran twice and committed two near-duplicate
suites (test_dependency_pins.py + test_requirements_versions.py) covering
the same pins. Keep the more robust one (operator-keyed specifier parsing
instead of next(iter(specifier)), which is order-fragile on multi-spec
requirements) and graft the three tests unique to the deleted file:
pyproject stale-comment guards, pipeline-lines-parseable, and the
wandb/typer click-comment consistency check.

Also:
- fix the copyright year (2026, not 2025)
- add tomli (python_version < 3.11) to common-tests.txt so the pyproject
  override tests actually RUN on the CI's Python 3.10 instead of
  silently skipping (tomllib is stdlib only from 3.11)
- add the missing trailing newline that failed the pre-commit
  end-of-file-fixer hook on the generated files

17 tests pass on Python 3.10 with the CI's -m 'not gpu' selection;
pre-commit (pinned ruff) clean.
